test(e2e): fix vmcompose for staging (#264)
Fixes a few things in e2e to support vmcompose to remote VMs.

task: none

---------

+++ test/e2e/app/cleanup.go
@@ -5,48 +5,22 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"runtime"
 
 	"github.com/omni-network/omni/lib/errors"
 	"github.com/omni-network/omni/lib/log"
 
-	e2e "github.com/cometbft/cometbft/test/e2e/pkg"
-	"github.com/cometbft/cometbft/test/e2e/pkg/exec"
 	"github.com/cometbft/cometbft/test/e2e/pkg/infra/docker"
 )
 
-// Cleanup removes the Docker Compose containers and testnet directory.
-func Cleanup(ctx context.Context, testnet *e2e.Testnet) error {
-	err := cleanupDocker(ctx)
-	if err != nil {
+// Cleanup removes the infra containers and testnet directory.
+func Cleanup(ctx context.Context, def Definition) error {
+	if err := def.Infra.Clean(ctx); err != nil {
+		return errors.Wrap(err, "cleaning infrastructure")
+	}
+
+	if err := cleanupDir(ctx, def.Testnet.Dir); err != nil {
 		return err
-	}
-	err = cleanupDir(ctx, testnet.Dir)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// cleanupDocker removes all E2E resources (with label e2e=True), regardless
-// of testnet.
-func cleanupDocker(ctx context.Context) error {
-	log.Info(ctx, "Removing docker containers and networks")
-
-	// GNU xargs requires the -r flag to not run when input is empty, macOS
-	// does this by default. Ugly, but works.
-	xargsR := `$(if [[ $OSTYPE == "linux-gnu"* ]]; then echo -n "-r"; fi)`
-
-	err := exec.Command(ctx, "bash", "-c", fmt.Sprintf(
-		"docker container ls -qa --filter label=e2e | xargs %v docker container rm -f", xargsR))
-	if err != nil {
-		return errors.Wrap(err, "remove docker containers")
-	}
-
-	err = exec.Command(ctx, "bash", "-c", fmt.Sprintf(
-		"docker network ls -q --filter label=e2e | xargs %v docker network rm", xargsR))
-	if err != nil {
-		return errors.Wrap(err, "remove docker networks")
 	}
 
 	return nil
@@ -70,18 +44,20 @@ func cleanupDir(ctx context.Context, dir string) error {
 	// On Linux, some local files in the volume will be owned by root since CometBFT
 	// runs as root inside the container, so we need to clean them up from within a
 	// container running as root too.
-	absDir, err := filepath.Abs(dir)
-	if err != nil {
-		return errors.Wrap(err, "abs dir")
-	}
-	err = docker.Exec(ctx, "run",
-		"--rm",             // Remove the container after it exits
-		"--entrypoint", "", // Clear the entrypoint so we can run a shell command
-		"-v", fmt.Sprintf("%v:/mount", absDir), // Mount the testnet dir into the container
-		"ethereum/client-go:latest",    // Use the latest geth image (which runs as root)
-		"sh", "-c", "rm -rf /mount/*/") // Remove all files in the mounted testnet dir
-	if err != nil {
-		return errors.Wrap(err, "docker exec rm mount")
+	if runtime.GOOS == "linux" {
+		absDir, err := filepath.Abs(dir)
+		if err != nil {
+			return errors.Wrap(err, "abs dir")
+		}
+		err = docker.Exec(ctx, "run",
+			"--rm",             // Remove the container after it exits
+			"--entrypoint", "", // Clear the entrypoint so we can run a shell command
+			"-v", fmt.Sprintf("%v:/mount", absDir), // Mount the testnet dir into the container
+			"ethereum/client-go:latest",    // Use the latest geth image (which runs as root)
+			"sh", "-c", "rm -rf /mount/*/") // Remove all files in the mounted testnet dir
+		if err != nil {
+			return errors.Wrap(err, "exec rm dir")
+		}
 	}
 
 	err = os.RemoveAll(dir)
